+++ src/app/app.routes.ts
@@ -2,25 +2,29 @@ import { Routes } from '@angular/router';
 
 import { taskRoutes } from './task.routes';
 import { authRoutes } from './auth.routes';
+import { AuthGuard } from './auth/guards';
 
 export const routes: Routes = [
   {
     path: '',
     pathMatch: 'full',
-    redirectTo: 'tasks'
+    redirectTo: 'tasks',
   },
   {
     path: 'auth',
-    loadComponent: () => import('./auth/layouts/auth-layout/auth-layout.component'),
-    children: [...authRoutes]
+    loadComponent: () =>
+      import('./auth/layouts/auth-layout/auth-layout.component'),
+    children: [...authRoutes],
   },
   {
     path: 'tasks',
-    loadComponent: () => import('./tasks/layouts/tasks-layout/tasks-layout.component'),
-    children: [...taskRoutes]
+    canMatch: [AuthGuard],
+    loadComponent: () =>
+      import('./tasks/layouts/tasks-layout/tasks-layout.component'),
+    children: [...taskRoutes],
   },
   {
     path: '**',
-    redirectTo: 'tasks'
-  }
+    redirectTo: 'tasks',
+  },
 ];