+++ src/app/auth/guards/auth.guard.ts
@@ -1,0 +1,36 @@
+import { inject } from '@angular/core';
+import { CanMatchFn, Route, Router, UrlSegment } from '@angular/router';
+
+import { firstValueFrom } from 'rxjs';
+
+import { AuthService } from '../services';
+
+export const AuthGuard: CanMatchFn = async (
+  route: Route,
+  segments: UrlSegment[]
+) => {
+  const authService = inject(AuthService);
+  const router = inject(Router);
+
+  const token = authService.token();
+
+  if (!token) {
+    return redirectToLogin(router);
+  }
+
+  const { isValidToken } = await firstValueFrom(
+    authService.validateToken({ token })
+  );
+
+  if (!isValidToken) {
+    authService.logout();
+    return redirectToLogin(router);
+  }
+
+  return true;
+};
+
+const redirectToLogin = (router: Router) => {
+  router.navigateByUrl('/auth/login');
+  return false;
+};